requirements.txt
opencv-python
ultralytics
numpy
pandas
matplotlib
scikit-learn
albumentations
PyYAML
